+++ UniversityManagementSystem.Domain/Interfaces/AcademicInterface/ICourseRepository.cs
@@ -5,13 +5,13 @@ namespace UniversityManagementSystem.Domain.Interfaces.AcademicInterface
 {
     public interface ICourseRepository
     {
-        Task<Course> GetCourseByCourseIdAsync(int courseId);
+        Task<Course> GetCourseByCourseIdAsync(ushort courseId);
         Task<Course> GetCourseByCourseCodeAsync(string code);
         Task<IEnumerable<Course>> GetAllCoursesAsync();
-        Task<IEnumerable<Course>> GetAllCoursesByFacultyIdAsync(int facultId);
-        Task<IEnumerable<Course>> GetAllCoursesByMajorIdAsync(int majorId);
+        Task<IEnumerable<Course>> GetAllCoursesByFacultyIdAsync(sbyte facultId);
+        Task<IEnumerable<Course>> GetAllCoursesByMajorIdAsync(ushort majorId);
         Task<int> AddCourseAsync(Course course);
-        Task<int> DeleteCourseAsync(int courseId);
+        Task<int> DeleteCourseAsync(ushort courseId);
         Task<int> UpdateCourseAsync(Course course);
     }
 }
