+++ UniversityManagementSystem.Domain/Interfaces/AlumniInterface/IAlumniService.cs
@@ -5,11 +5,11 @@ namespace UniversityManagementSystem.Domain.Interfaces.AlumniInterface
 {
     public interface IAlumniService
     {
-        Task<Alumni> GetAlumniByIdAsync(int id);
+        Task<Alumni> GetAlumniByIdAsync(uint id);
         Task<IEnumerable<Alumni>> GetAllAlumniAsync();
-        Task<IEnumerable<Alumni>> GetAllAlumniByYearAsync(int year);
-        Task<IEnumerable<Alumni>> GetAllAlumniByMajorIdAsync(int id);
-        Task<IEnumerable<Alumni>> GetAllAlumniByFacultyIdAsync(int id);
+        Task<IEnumerable<Alumni>> GetAllAlumniByYearAsync(ushort year);
+        Task<IEnumerable<Alumni>> GetAllAlumniByMajorIdAsync(ushort id);
+        Task<IEnumerable<Alumni>> GetAllAlumniByFacultyIdAsync(sbyte id);
         Task<IEnumerable<Alumni>> GetAllUndergraduateAlumniAsync();
         Task<IEnumerable<Alumni>> GetAllGraduateAlumniAsync();
         Task<IEnumerable<Alumni>> GetAllDoctoralAlumniAsync();
@@ -17,7 +17,7 @@ namespace UniversityManagementSystem.Domain.Interfaces.AlumniInterface
         Task<int> AddCurrentYearAlumniAsync(IEnumerable<Alumni> alumniList);
         Task<int> UpdateAlumniAsync(Alumni alumni);
         Task<int> BatchUpdateAlumniAsync(IEnumerable<Alumni> alumniList);
-        Task<int> DeleteAlumniAsync(int id);
-        Task<int> BatchDeleteAlumniAsync(IEnumerable<int> idList);
+        Task<int> DeleteAlumniAsync(uint id);
+        Task<int> BatchDeleteAlumniAsync(IEnumerable<uint> idList);
     }
 }
